# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AgentCordon API
  version: 1.6.1
  description: |
    AgentCordon is a self-hostable Agentic Identity Provider (IdP) that provides
    secure, auditable, policy-controlled token brokerage for autonomous agents.
    Agents do not directly hold long-lived credentials; every issuance and use
    is observable via the audit log.
  contact:
    name: AgentCordon
  license:
    name: MIT
servers:
  - url: http://localhost:3140
    description: Local development server
...
security:
  - SessionCookie: []
  - BearerJWT: []
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: agtcrdn_session
      description: |
        Session cookie set after user login. Every unsafe method authenticated
        this way is also CSRF-protected by a double-submit rule: the request
        must carry an `X-CSRF-Token` header whose value matches the readable
        `agtcrdn_csrf` cookie, or it is refused with
        `403 csrf_validation_failed`.
    BearerJWT:
      type: http
      scheme: bearer
      description: |
        An **opaque** OAuth 2.0 access token issued to a workspace by
        `POST /api/v1/oauth/token`. Despite the scheme name this is not a JWT
        and carries no claims a client can read: the server stores only its
        SHA-256 hash and resolves the workspace by token id. The ES256 JWT
        issuer, its JWKS endpoint, and the permissions-token endpoint were
        removed.
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok, error]
        data:
          description: Response payload (present when status is ok)
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    Vault:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: >-
            Display name, 1-100 characters. Unique among the vaults one user
            owns; free across owners, so two users may each have a `team`.
        owner_user_id:
          type: string
          format: uuid
          nullable: true
          description: The owner, or null for the system default vault.
        is_default:
          type: boolean
          description: >-
            true for the one vault (00000000-0000-0000-0000-000000000001) every
            user may put a credential in and nobody may rename, share or delete.
        shared_by:
          type: string
          description: >-
            Present only when the caller sees this vault through a share: the
            username of whoever shared it.
        permission:
          type: string
          enum: [read]
          description: Present only on a vault shared with the caller.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    VaultShare:
      type: object
      properties:
        id:
          type: string
        vault_id:
          type: string
          format: uuid
        shared_with_user_id:
          type: string
          format: uuid
        permission_level:
          type: string
          enum: [read]
          description: >-
            read is the only level the authorization model keeps. write and
            admin are refused with 400 rather than stored as a promise nothing
            keeps.
        shared_by_user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/vaults:
    get:
      tags: [Vaults]
      summary: List vaults
      description: |
        The vaults the caller can see: the ones they own, the system default,
        the ones shared with them (each naming `shared_by` and `permission`),
        and -- for a holder of `manage_vaults` -- every vault on the install.
        A workspace owns no vaults and is party to no shares, so it gets an
        empty list.
      security:
        - BearerJWT: []
        - SessionCookie: []
      responses:
        '200':
          description: List of vaults
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Vault'
    post:
      tags: [Vaults]
      summary: Create a vault
      description: |
        Creates a vault owned by the caller. Anyone the policy engine lets
        create credentials may create somewhere to put them. A name is unique
        among the caller's own vaults and free across owners.
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 100
      responses:
        '200':
          description: Vault created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Vault'
        '400':
          description: Name is empty or longer than 100 characters
        '403':
          description: Workspaces cannot create vaults
        '409':
          description: The caller already owns a vault with this name
  /api/v1/vaults/{id}:
    patch:
      tags: [Vaults]
      summary: Rename a vault
      description: |
        The vault's owner, or root. The name is a label, so this changes
        nothing else.
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 100
      responses:
        '200':
          description: Vault renamed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Vault'
        '403':
          description: >-
            Not the owner, or the system default vault, which cannot be
            renamed
        '404':
          description: Vault not found
        '409':
          description: The owner already has another vault with this name
    delete:
      tags: [Vaults]
      summary: Delete an empty vault
      description: |
        The vault's owner, or root. A vault that still holds credentials is a
        conflict, not a cascade: deleting it would take the credentials with
        it, which is never what "remove this grouping" meant. Move or delete
        them first.
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vault deleted
        '403':
          description: >-
            Not the owner, or the system default vault, which cannot be deleted
        '404':
          description: Vault not found
        '409':
          description: The vault still holds credentials
  /api/v1/vaults/{id}/credentials:
    get:
      tags: [Vaults]
      summary: List credentials in a vault
      description: |
        Credential summaries for the vault. Non-root users only see
        credentials they created, in vaults they own, or in vaults shared with
        them.
      security:
        - BearerJWT: []
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Credentials in vault
  /api/v1/vaults/{id}/shares:
    get:
      tags: [Vaults]
      summary: List vault shares
      description: |
        The share list names every user this vault reaches, so a signed-in
        stranger is refused rather than handed the roster: the caller must own
        the vault, be root, or hold `manage_vaults`.
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of vault shares
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/VaultShare'
        '403':
          description: Not the owner and not a manage_vaults holder
        '404':
          description: Vault not found

    post:
      tags: [Vaults]
      summary: Share a vault with a user
      description: |
        Granting is the vault owner's act and nobody else's, whatever role they
        hold. Holding `manage_vaults` is deliberately **not** enough: it reads
        and revokes, it does not give away another user's credentials.

        `read` is the only level the authorization model supports; `write` and
        `admin` are refused with 400 rather than stored as a promise nothing
        keeps. A read share makes the vault's credentials visible in the
        recipient's list and detail, and nothing more.
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
                  format: uuid
                permission:
                  type: string
                  enum: [read]
                  default: read
      responses:
        '200':
          description: Vault shared
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/VaultShare'
        '400':
          description: permission was 'write' or 'admin', or user_id is malformed
        '403':
          description: >-
            Not the owner, or the system default vault, which cannot be shared
        '404':
          description: Target user not found
  /api/v1/vaults/{id}/shares/{user_id}:
    delete:
      tags: [Vaults]
      summary: Revoke vault sharing
      description: |
        The vault's owner, root, or a holder of `manage_vaults` -- unlike
        granting, revoking is also an admin power, so someone can cut off a
        share they did not make.
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Share revoked
        '403':
          description: Not the owner and not a manage_vaults holder
        '404':
          description: Vault share not found

  # ============================================================================
  # OIDC
  # ============================================================================
...